requirements.txt
pillow~=10.2.0
aiosqlite~=0.19.0
aiogram~=3.3.0
python-dotenv~=1.0.0
SQLAlchemy~=2.0.25